Renamed rpo and updated README title accordingly
The server component is not "a website". Its main function will actually be the queue management and tcp server that provides the gateway(s) with the messages to be broadcasted.

It will (if time provides) also involve a web interface, but the "website" is usually used to mention content and/or static web pages, not web applications or web interfaces to other systems.